feat: 調整部分功能放到 Backend 下的 Controller
說明：
    1. 非對外公開的 API 放到 Backend 下的 Controller。

調整項目：
    1. */CharacterController.php:
        - 非對外公開的 API 放到 Backend 下的 Controller。
    2. api.php:
        - 將專武清單 API 設為對外公開。
    3. web.php:
        - 依據 1. 的調整修正 Controller 的命名空間。
